+++ src/exports/fields.ts
@@ -1,4 +1,5 @@
 // Field configurations
-export { createEmailContentField, emailSafeFeatures, createEmailSafeFeatures } from '../fields/emailContent'
+export { createEmailContentField, emailSafeFeatures, createEmailSafeFeatures, createEmailLexicalEditor } from '../fields/emailContent'
 export { createBroadcastInlinePreviewField } from '../fields/broadcastInlinePreview'
 export { createNewsletterSchedulingFields } from '../fields/newsletterScheduling'
+export { createEmailSafeBlocks, validateEmailBlocks } from '../utils/blockValidation'